Moved example agents to seperate file and added strong and weak opponent
