+++ teamprojekt_competition_server/client/main.py
@@ -1,9 +1,7 @@
 """main class with dummy agent for testing"""
-import logging as log
-import numpy as np
+from .example_agents import Strong_Hockey_Agent
 
-from .agent import COMPAgent
-import random
+import logging as log
 
 log.basicConfig(level=log.DEBUG)
 
@@ -11,39 +9,5 @@ log.basicConfig(level=log.DEBUG)
 # run with "python -m teamprojekt_competition_server.client.main"
 
 if __name__ == "__main__":
-
-    class Laserhockey_Agent(COMPAgent):
-        """Dummy Agent for testing"""
-
-        def step(self, obv: list[float]) -> list[float]:
-            """dummy step function
-
-            Args:
-                obv (list[float]): observation
-
-            Returns:
-                list(float): action
-            """
-
-            return np.random.uniform(-1, 1, 4).tolist()
-            # Action Space Box(-1.0, 1.0, (3,), float32)
-
-    class Rock_Paper_Scissors_Agent(COMPAgent):
-        """Dummy Agent for testing"""
-
-        def step(self, obv: list[float]) -> list[float]:
-            """dummy step function
-
-            Args:
-                obv (list[float]): environment
-
-            Returns:
-                list(float): action
-            """
-            # return float(input(f"Observation: {obv} | Enter a move: "))
-            print(random.choice([0.0, 1.0, 2.0]))
-            return [random.choice([0.0, 1.0, 2.0])]
-            # Action Space Box(-2.0, 2.0, (1,), float32)
-
-    agent = Laserhockey_Agent()  # Rock_Paper_Scissors_Agent()
+    agent = Strong_Hockey_Agent()  
     agent.run("HelloWorldToken")
